+++ traffic_ops/testing/api/v3/regions_test.go
@@ -25,6 +25,7 @@ func TestRegions(t *testing.T) {
 	WithObjs(t, []TCObj{Parameters, Divisions, Regions}, func() {
 		UpdateTestRegions(t)
 		GetTestRegions(t)
+		DeleteTestRegionsByName(t)
 	})
 }
 
@@ -82,6 +83,26 @@ func GetTestRegions(t *testing.T) {
 			t.Errorf("cannot GET Region by region: %v - %v", err, resp)
 		}
 	}
+}
+
+func DeleteTestRegionsByName(t *testing.T) {
+	for _, region := range testData.Regions {
+		delResp, _, err := TOSession.DeleteRegion(nil, &region.Name)
+		if err != nil {
+			t.Errorf("cannot DELETE Region by name: %v - %v", err, delResp)
+		}
+
+		// Retrieve the Region to see if it got deleted
+		regionResp, _, err := TOSession.GetRegionByName(region.Name)
+		if err != nil {
+			t.Errorf("error deleting Region region: %s", err.Error())
+		}
+		if len(regionResp) > 0 {
+			t.Errorf("expected Region : %s to be deleted", region.Name)
+		}
+	}
+
+	CreateTestRegions(t)
 }
 
 func DeleteTestRegions(t *testing.T) {